bloomfilter: add (de)serialise functions, for better abstraction
At the moment, the bloomfilter sub-library does not itself support
serialisation, and this is handled by the lsm-tree lib directly. It does
so by being overly familiar with the internal representation of the
bloomfilter lib.

We still want the file format to be determined by the lsm-tree library,
but we would like a better abstraction boundary between the bloomfilter
and lsm-tree libraries.

So we introduce functions to convert the bloom filters to/from a lower
level representation as a BloomSize and a (Mutable)ByteArray for the
bits.