bloomfilter: switch range reduction from division to multiplication
In the Classic implementation. The Blocked one uses this already.

The classic algorithm is the remainder after division. But division is
slow, especially for 64bit numbers. There's a faster method based on
multiplication.

This changes the format version, since it changes the mapping from
element hash to the bit selected.