bloomfilter: add bloomfilter-fpr-calc and gnuplot script
The program produces output for gnuplot, and the gnuplot script plots
the calculated and actual FPR vs bits-per-key for both the classic and
the blocked implementations.

We use 2..20 bits for classic, as that gets us down to below an FPR of
1e-4, while for the blocked one we need to use 2..24 bits to get to
approximately the same FPR.

This graph output (checked in) is also a great test of the FPR to bits
code for both implementations, and of the FPR of each impl in general.