bloomfilter: no longer export Bloom filter representations
All the functions that need it are now in the Internal module.

Hide the internal modules, using 'other-modules', rather than
'exposed-modules'.

The last code change needed was to remove use of the internal invariant
in in the deserialisation test. This is no longer needed because the
deserialisation code being testeeed itself no longer has access to the
internal representation. It now uses a proper API for deserialisation,
which itself uses the invariant checking.