fpr-calc: rename main output to fpr-vs-bits
In preparation for also adding an analysis for fpr-vs-fpr.

The fpr-vs-bits plots bits per key on the X axis, and calculated and
actual FPR on the Y axis.

A for-vs-fpr will plot _requested_ FPR on the X axis and actual FPR on
the Y axis. For our impls this is a sanity check: it should look like a
straight line. But for the original bloomfilter package it is all over
the place (due to using power of two sized filters).